WF-DataImporter Ver.1.0 Release Changes
In preparation to release this one to the people, made a few changes.

Old PowerShell Version:
Moves the old PowerShell variant to 'WFDI-PowerShell' along with the 'bin' folder as its not used normally anymore.

Comments Galore:
Added comments all through the code to help anyone out in the future.

File Cleanup:
As we have moved our work on damage calculation over to our main TenZeroGG git, I cleared out that folder and also cleared up the 'Custom' folder under 'JSON'. I am leaving the custom files as they are of course not our data really, even if we made them.